Safari-609.1.11.1/LayoutTests/http/tests/navigation/anchor-blank-target-implies-rel-noopener-expected.txt

CONSOLE MESSAGE: line 6: PASS: New window should not have an opener
Tests that a new window opened via target=_blank does not have an opener
